feat(map): Add additionnal menu to left atlases
