+++ vue/src/services/qgisMap/QgisMapService.ts
@@ -13,9 +13,14 @@ export class QgisMapService {
       .then((response) => response.data['hydra:member'][0])
   }
 
-  static async post(qgisMap: QgisMap): Promise<QgisMap> {
-    qgisMap = await this.handleQgisProjectUpload(qgisMap)
-    return await apiClient.post('/api/qgis_maps', qgisMap).then((response) => response.data)
+  static async post(qgisMap: QgisMap): Promise<QgisMap | null> {
+    try {
+      qgisMap = await this.handleQgisProjectUpload(qgisMap)
+      return await apiClient.post('/api/qgis_maps', qgisMap).then((response) => response.data)
+    } catch (error) {
+      console.error(error)
+      return null
+    }
   }
 
   static async patch(qgisMap: QgisMap): Promise<QgisMap> {
@@ -33,13 +38,17 @@ export class QgisMapService {
    * Upload all files in the object and replace them with the media object iri
    */
   static handleQgisProjectUpload = async (object: any) => {
-    const uploadPromises = Object.keys(object).map(async (fieldName) => {
-      if (object[fieldName] instanceof File) {
-        const response = await QgisProjectUploader.uploadFile(object[fieldName])
-        object[fieldName] = response['@id']
-      }
-    })
-    await Promise.all(uploadPromises)
-    return object
+    try {
+      const uploadPromises = Object.keys(object).map(async (fieldName) => {
+        if (object[fieldName] instanceof File) {
+          const response = await QgisProjectUploader.uploadFile(object[fieldName])
+          object[fieldName] = response['@id']
+        }
+      })
+      await Promise.all(uploadPromises)
+      return object
+    } catch (e) {
+      console.error(e)
+    }
   }
 }
